move code from the snapshot_service into the vit-servicing-station-lib, remove snapshot_service crate
