--- snapshot-service/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "snapshot-service"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-snapshot-lib = { git = "https://github.com/input-output-hk/catalyst-toolbox", branch = "main" }
-notify = "=5.0.0-pre.11"
-serde = { version = "1.0.136", features = ["derive"] }
-serde_json = "1.0.79"
-tracing = "0.1.34"
-thiserror = "1.0.30"
-warp = "0.3.2"
-chain-ser = { git = "https://github.com/input-output-hk/chain-libs", branch = "master" }
-jormungandr-lib = { git = "https://github.com/input-output-hk/jormungandr", branch = "master" }
-sled = "0.34.7"
-tokio = "1.18.0"
-time = "0.3"
