+++ vit-servicing-station-lib/Cargo.toml
@@ -2,7 +2,7 @@
 name = "vit-servicing-station-lib"
 version = "0.3.4-dev"
 authors = ["danielsanchezq <daniel.sanchez@iohk.io>"]
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -35,7 +35,6 @@ eccoxide = { git = "https://github.com/eugene-babichenko/eccoxide.git", branch =
 http-zipkin = "0.3.0"
 notify = "=5.0.0-pre.11"
 sled = "0.34.7"
-snapshot-service = { path = "../snapshot-service" }
 
 # This solves building on windows when sqlite3lib is not installed or missing in the `$PATH`
 # as it happens with the github actions pipeline associated to this project.
