+++ vit-servicing-station-server/Cargo.toml
@@ -15,4 +15,3 @@ tracing-futures = "0.2.4"
 tracing-subscriber = { version = "0.3", features=["fmt"] }
 tracing-appender = "0.2"
 vit-servicing-station-lib = { path = "../vit-servicing-station-lib" }
-snapshot-service = { path = "../snapshot-service" }
