+++ vit-servicing-station-tests/Cargo.toml
@@ -23,7 +23,6 @@ assert_fs = "1.0.0"
 assert_cmd = "2"
 hyper = "0.14.2"
 vit-servicing-station-lib = {path= "../vit-servicing-station-lib"}
-snapshot-service = {path= "../snapshot-service"}
 lazy_static = "1.4"
 rand = "0.7"
 url = "2.2"
